Renamed tree_math -> tree_util; moved HashableLeaf to tree_util
