+++ src/tree_util.rs
@@ -1,3 +1,11 @@
+use digest::Digest;
+
+/// The domain separator used for calculating leaf hashes
+const LEAF_HASH_PREFIX: &[u8] = &[0x00];
+
+/// The domain separator used for calculating parent hashes
+const PARENT_HASH_PREFIX: &[u8] = &[0x01];
+
 // We make opaque types for leaf and internal node indices so that we don't accidentally confuse
 // them in the math
 
@@ -37,6 +45,60 @@ impl InternalIdx {
     pub(crate) fn as_usize(&self) -> Option<usize> {
         self.0.try_into().ok()
     }
+}
+
+/// Represents a leaf that can be included in a Merkle tree. This only requires that the leaf have a
+/// unique hash representation.
+pub trait HashableLeaf {
+    fn hash<H: digest::Update>(&self, hasher: &mut H);
+}
+
+// Blanket hasher impl for anything that resembles a bag of bytes
+impl<T: AsRef<[u8]>> HashableLeaf for T {
+    fn hash<H: digest::Update>(&self, hasher: &mut H) {
+        hasher.update(self.as_ref())
+    }
+}
+
+/// A hasher that prepends the leaf-hash prefix
+struct LeafHasher<H: Digest>(H);
+
+impl<H: Digest> LeafHasher<H> {
+    fn new() -> Self {
+        LeafHasher(H::new_with_prefix(LEAF_HASH_PREFIX))
+    }
+
+    fn finalize(self) -> digest::Output<H> {
+        self.0.finalize()
+    }
+}
+
+impl<H: Digest> digest::Update for LeafHasher<H> {
+    fn update(&mut self, data: &[u8]) {
+        self.0.update(data)
+    }
+}
+
+/// Computes the hash of the given leaf's canonical byte representation
+pub(crate) fn leaf_hash<H, L>(leaf: &L) -> digest::Output<H>
+where
+    H: Digest,
+    L: HashableLeaf,
+{
+    let mut hasher = LeafHasher::<H>::new();
+    leaf.hash(&mut hasher);
+    hasher.finalize()
+}
+
+/// Computes the parent of the two given subtrees. This is `H(0x01 || left || right)`.
+pub(crate) fn parent_hash<H: Digest>(
+    left: &digest::Output<H>,
+    right: &digest::Output<H>,
+) -> digest::Output<H> {
+    let mut hasher = H::new_with_prefix(PARENT_HASH_PREFIX);
+    hasher.update(left);
+    hasher.update(right);
+    hasher.finalize()
 }
 
 //
